Improve player physics
- Grass block can be created with “E” and deleted with “Q”
- Add trees
- Add random number generator for perlin noise
- Add a seed variable to the world class
- Optimized the rendering of the display lists
- Support for translucent blocks

+++ src/main/java/com/continuum/RenderObject.java
@@ -25,9 +25,9 @@ public  abstract class RenderObject
 	/**
 	 * Update method
 	 *
-	 * @param delta
+	 * @param long
 	 */
-	public void update(int delta) {}
+	public void update(long delta) {}
 
 	/**
 	 * Get object position.
